+++ sos-webservices/sos-webservices-joc/src/main/java/com/sos/joc/tags/group/impl/GroupsModifyImpl.java
@@ -154,7 +154,7 @@ public class GroupsModifyImpl extends JOCResourceImpl implements ITagsModify {
                 if (dbItem == null) {
                     continue;
                 }
-                if (ordering != dbItem.getOrdering()) {
+                if (ordering != dbItem.getOrdering().intValue()) {
                     dbItem.setOrdering(ordering);
                     dbLayer.getSession().update(dbItem);
                 }
@@ -162,7 +162,7 @@ public class GroupsModifyImpl extends JOCResourceImpl implements ITagsModify {
             }
             for (DBItemInventoryTagGroup dbItem : mappedByName.values().stream().sorted(Comparator.comparingInt(DBItemInventoryTagGroup::getOrdering))
                     .collect(Collectors.toCollection(LinkedList::new))) {
-                if (ordering != dbItem.getOrdering()) {
+                if (ordering != dbItem.getOrdering().intValue()) {
                     dbItem.setOrdering(ordering);
                     dbLayer.getSession().update(dbItem);
                 }
